+++ src/main/java/seedu/recipe/MainApp.java
@@ -28,8 +28,8 @@ import seedu.recipe.storage.RecipeBookStorage;
 import seedu.recipe.storage.Storage;
 import seedu.recipe.storage.StorageManager;
 import seedu.recipe.storage.UserPrefsStorage;
-import seedu.recipe.logic.commands.ui.Ui;
-import seedu.recipe.logic.commands.ui.UiManager;
+import seedu.recipe.ui.Ui;
+import seedu.recipe.ui.UiManager;
 
 /**
  * Runs the application.
