+++ src/test/java/seedu/recipe/logic/commands/SubCommandTest.java
@@ -1,0 +1,4 @@
+package seedu.recipe.logic.commands;
+
+public class SubCommandTest {
+}
